+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Fixed
+- Prune was not working with old TF versions because of the `depends_on`
+  ([Issue #144](https://github.com/cycloidio/inframap/pull/144))
+
 ## [0.6.6] _2021-06-10_
 
 ### Changed